ai-agent-ms/requirements.txt
flask
google-genai
pypdf2
numpy
pandas
scikit-learn
tenacity 